requirements.txt
apollox==0.0.2
binance==0.3
ftx==1.2.0
kucoin==0.1
kucoin_futures_python==1.0.6
pymongo==4.1.1
python_kucoin==2.1.3
requests==2.28.0
